requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
nltk==3.8.1
datasets==2.14.5
joblib==1.3.2
python-multipart==0.0.6
jinja2==3.1.2
websockets==11.0.3